more changes, visual test runner refactor; cpp2023
  1. CLAUDE.md - Added make build-tests command and added a fun note about cookies
  2. CMakeLists.txt - Updated C++ standard from C++17 to C++23, added new collision/adhesion
  calculators
  3. Makefile - Added build-tests target to build tests without running them
  4. ui_overview.md - Added documentation for UI control positions (Y coordinates)
  5. visual_test_framework.md - Major updates emphasizing restart functionality and step mode as
  defaults
  6. Cell.cpp & CellB.cpp - Fixed opacity calculations with explicit double casting
  7. SimulatorUI.cpp - Fixed style selector casting for button states
  8. WorldB.cpp - Refactored to use new calculator classes for collisions and adhesion
  9. Multiple new calculator files - Modularized collision and adhesion logic into separate
  classes

  The main changes involve refactoring collision/adhesion logic into dedicated calculator classes
  and updating the visual test framework documentation to promote better testing practices.

+++ test-lvgl/src/Cell.cpp
@@ -168,8 +168,8 @@ void Cell::drawNormal(lv_obj_t* /* parent */, uint32_t /* x */, uint32_t /* y */
     lv_color_t brown = lv_color_hex(0x8B4513); // Saddle brown color
 
     // Calculate opacity based on dirt amount (0.0 to 1.0)
-    lv_opa_t opacity_dirt = static_cast<lv_opa_t>(dirt * LV_OPA_COVER);
-    lv_opa_t opacity_water = static_cast<lv_opa_t>(water * LV_OPA_COVER);
+    lv_opa_t opacity_dirt = static_cast<lv_opa_t>(dirt * static_cast<double>(LV_OPA_COVER));
+    lv_opa_t opacity_water = static_cast<lv_opa_t>(water * static_cast<double>(LV_OPA_COVER));
 
     lv_layer_t layer;
     lv_canvas_init_layer(canvas, &layer);
@@ -215,8 +215,8 @@ void Cell::drawDebug(lv_obj_t* /* parent */, uint32_t /* x */, uint32_t /* y */)
     lv_color_t brown = lv_color_hex(0x8B4513); // Saddle brown color
 
     // Calculate opacity based on dirt amount (0.0 to 1.0)
-    lv_opa_t opacity_dirt = static_cast<lv_opa_t>(dirt * LV_OPA_COVER);
-    lv_opa_t opacity_water = static_cast<lv_opa_t>(water * LV_OPA_COVER);
+    lv_opa_t opacity_dirt = static_cast<lv_opa_t>(dirt * static_cast<double>(LV_OPA_COVER));
+    lv_opa_t opacity_water = static_cast<lv_opa_t>(water * static_cast<double>(LV_OPA_COVER));
 
     lv_layer_t layer;
     lv_canvas_init_layer(canvas, &layer);
