+++ test-lvgl/src/CellB.cpp
@@ -628,7 +628,7 @@ void CellB::drawNormal(lv_obj_t* parent, uint32_t x, uint32_t y)
         }
 
         // Calculate opacity based on fill ratio
-        lv_opa_t opacity = static_cast<lv_opa_t>(fill_ratio_ * LV_OPA_COVER);
+        lv_opa_t opacity = static_cast<lv_opa_t>(fill_ratio_ * static_cast<double>(LV_OPA_COVER));
 
         // Draw material layer with enhanced border to match debug mode
         lv_draw_rect_dsc_t rect_dsc;
@@ -724,7 +724,7 @@ void CellB::drawDebug(lv_obj_t* parent, uint32_t x, uint32_t y)
 
         if (material_type_ != MaterialType::AIR) {
             // Calculate opacity based on fill ratio
-            lv_opa_t opacity = static_cast<lv_opa_t>(fill_ratio_ * LV_OPA_COVER);
+            lv_opa_t opacity = static_cast<lv_opa_t>(fill_ratio_ * static_cast<double>(LV_OPA_COVER));
 
             // Draw material layer with enhanced border for debug
             lv_draw_rect_dsc_t rect_dsc;
